chore(desktop): split settings cleanup files
