+++ scripts/max_lines_allowlist.txt
@@ -48,6 +48,7 @@ anyharness/sdk-react/src/hooks/sessions.ts
 desktop/src/components/ui/icons.tsx
 desktop/src/hooks/sessions/use-session-creation-actions.ts
 desktop/src/hooks/sessions/use-session-runtime-actions.ts
+desktop/src/hooks/sessions/lifecycle/use-session-stream-flush.ts
 desktop/src/hooks/workspaces/tabs/use-workspace-shell-activation.ts
 desktop/src/hooks/workspaces/tabs/use-workspace-header-subagent-hierarchy.ts
 desktop/src/hooks/workspaces/use-workspace-bootstrap-actions.ts
